Fixed router and redux store with login and logout
